EQ4-28-Merging 2 missing files from backend
