+++ backend/src/main/java/com/team4/backend/controller/FileMetaDataController.java
@@ -3,21 +3,35 @@ package com.team4.backend.controller;
 import com.team4.backend.dto.FileMetaDataInternshipManagerViewDto;
 import com.team4.backend.mapping.FileMetaDataMapper;
 import com.team4.backend.service.FileMetaDataService;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.http.codec.multipart.FilePart;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
+import java.security.Principal;
+
 @RestController
-@RequestMapping("/fileMetaData")
+@RequestMapping("/file")
 public class FileMetaDataController {
 
-    private final FileMetaDataService fileMetaDataService;
+    @Autowired
+    FileMetaDataService fileMetaDataService;
 
-    public FileMetaDataController(FileMetaDataService fileMetaDataService) {
-        this.fileMetaDataService = fileMetaDataService;
+    protected String getLoggedUserName(Principal loggedUser) {
+        if (loggedUser == null) {
+            return "";
+        }
+        return loggedUser.getName();
+    }
+
+    @PostMapping
+    @PreAuthorize("hasAnyAuthority('STUDENT')")
+    public Mono<ResponseEntity<Void>> uploadFile(@RequestPart("filename") String filename, @RequestPart("type") String type, @RequestPart("mimeType") String mimeType, @RequestPart("file") Mono<FilePart> filePartMono, Principal loggedUser) {
+        return fileMetaDataService.uploadFile(filename, type, mimeType, filePartMono, getLoggedUserName(loggedUser));
     }
 
     @GetMapping("/countAllInvalidCvNotSeen")
@@ -42,3 +56,22 @@ public class FileMetaDataController {
     }
 
 }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
